+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add instructions for runing under Okapi.
 * Rename from `mod-app-store` to `mod-app-manager`
 * When requests provide Okapi details (url, tenant, token) use these instead of configured
+* Add `env` logging category
 
 ## [1.1.0](https://github.com/MikeTaylor/mod-app-store/tree/v1.1.0) (2022-07-18)